requirements.txt
# Путь: /youtube_automation_bot/requirements.txt
# Описание: Зависимости Python для проекта

# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
celery==5.3.4
redis==5.0.1
python-dotenv==1.0.0

# Telegram Bot
aiogram==3.3.0

# YouTube & Video Processing
yt-dlp==2024.1.0
ffmpeg-python==0.2.0

# AI Services
anthropic==0.8.1
openai-whisper==20231117
torch==2.1.2
torchaudio==2.1.2

# Yandex Services
yadisk==3.1.0
aiohttp==3.9.1

# Database
sqlalchemy==2.0.23
alembic==1.13.0

# Utils
pydantic==2.5.2
python-multipart==0.0.6
numpy==1.24.3
psutil==5.9.6

# Monitoring
flower==2.0.1  # Celery monitoring